requirements.txt
torch>=2.0.0
transformers>=4.30.0
onnx>=1.14.0
onnxruntime>=1.15.0
streamlit>=1.24.0
psutil>=5.9.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0
nvidia-ml-py3>=7.352.0
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0
